fix: update test name
